move preset and superset.py into its own directory
